+++ lib/oboe/frameworks/rails/inst/action_controller.rb
@@ -3,7 +3,82 @@
 
 module Oboe
   module Inst
-    module Rails3ActionController
+    #
+    # RailsBase
+    #
+    # This module contains the instrumentation code common to
+    # many Rails versions.
+    #
+    module RailsBase
+      #
+      # has_handler?
+      #
+      # Determins if <tt>exception</tt> has a registered
+      # handler via <tt>rescue_from</tt>
+      #
+      def has_handler?(exception)
+        # Don't log exceptions if they have a rescue handler set
+        has_handler = false
+        rescue_handlers.detect { | klass_name, handler |
+          # Rescue handlers can be specified as strings or constant names
+          klass = self.class.const_get(klass_name) rescue nil
+          klass ||= klass_name.constantize rescue nil
+          has_handler = exception.is_a?(klass) if klass
+        }
+        has_handler
+      rescue => e
+        Oboe.logger.debug "[oboe/debug] Error searching Rails handlers: #{e.message}"
+        return false
+      end
+
+      #
+      # log_rails_error?
+      #
+      # Determins whether we should log a raised exception to the
+      # TraceView dashboard.  This is determined by whether the exception
+      # has a rescue handler setup and the value of
+      # Oboe::Config[:report_rescued_errors]
+      #
+      def log_rails_error?(exception)
+        # As it's perculating up through the layers...  make sure that
+        # we only report it once.
+        return false if exception.instance_variable_get(:@oboe_logged)
+
+        has_handler = has_handler?(exception)
+
+        if !has_handler || (has_handler && Oboe::Config[:report_rescued_errors])
+          return true
+        end
+        false
+      end
+
+      #
+      # render_with_oboe
+      #
+      # Our render wrapper that just times and conditionally
+      # reports raised exceptions
+      #
+      def render_with_oboe(*args)
+        Oboe::API.log_entry('actionview')
+        render_without_oboe(*args)
+
+      rescue Exception => e
+        Oboe::API.log_exception(nil, e) if log_rails_error?(e)
+        raise
+      ensure
+        Oboe::API.log_exit('actionview')
+      end
+    end
+
+    #
+    # ActionController3
+    #
+    # This modules contains the instrumentation code specific
+    # to Rails v3
+    #
+    module ActionController3
+      include ::Oboe::Inst::RailsBase
+
       def self.included(base)
         base.class_eval do
           alias_method_chain :process, :oboe
@@ -13,9 +88,14 @@ module Oboe
       end
 
       def process_with_oboe(*args)
-        Oboe::API.trace('rails', {}) do
-          process_without_oboe *args
-        end
+        Oboe::API.log_entry('rails')
+        process_without_oboe *args
+
+      rescue Exception => e
+        Oboe::API.log_exception(nil, e) if log_rails_error?(e)
+        raise
+      ensure
+        Oboe::API.log_exit('rails')
       end
 
       def process_action_with_oboe(*args)
@@ -31,15 +111,17 @@ module Oboe
         Oboe::API.log(nil, 'info', report_kvs)
         raise
       end
-
-      def render_with_oboe(*args)
-        Oboe::API.trace('actionview', {}) do
-          render_without_oboe *args
-        end
-      end
     end
 
-    module Rails4ActionController
+    #
+    # ActionController4
+    #
+    # This modules contains the instrumentation code specific
+    # to Rails v4
+    #
+    module ActionController4
+      include ::Oboe::Inst::RailsBase
+
       def self.included(base)
         base.class_eval do
           alias_method_chain :process_action, :oboe
@@ -55,15 +137,15 @@ module Oboe
           :Controller   => self.class.name,
           :Action       => self.action_name,
         }
-        Oboe::API.trace('rails', report_kvs) do
-          process_action_without_oboe(method_name, *args)
-        end
-      end
 
-      def render_with_oboe(*args)
-        Oboe::API.trace('actionview', {}) do
-          render_without_oboe(*args)
-        end
+        Oboe::API.log_entry('rails')
+        process_action_without_oboe(method_name, *args)
+
+      rescue Exception => e
+        Oboe::API.log_exception(nil, e) if log_rails_error?(e)
+        raise
+      ensure
+        Oboe::API.log_exit('rails')
       end
     end
   end
@@ -73,27 +155,34 @@ if defined?(ActionController::Base) && Oboe::Config[:action_controller][:enabled
   if ::Rails::VERSION::MAJOR == 4
 
     class ActionController::Base
-      include Oboe::Inst::Rails4ActionController
+      include Oboe::Inst::ActionController4
     end
 
   elsif ::Rails::VERSION::MAJOR == 3
 
     class ActionController::Base
-      include Oboe::Inst::Rails3ActionController
+      include Oboe::Inst::ActionController3
     end
 
   elsif ::Rails::VERSION::MAJOR == 2
 
     ActionController::Base.class_eval do
+      include ::Oboe::Inst::RailsBase
+
       alias :perform_action_without_oboe :perform_action
       alias :rescue_action_without_oboe :rescue_action
       alias :process_without_oboe :process
       alias :render_without_oboe :render
 
       def process(*args)
-        Oboe::API.trace('rails', {}) do
-          process_without_oboe(*args)
-        end
+        Oboe::API.log_entry('rails')
+        process_without_oboe(*args)
+
+      rescue Exception => e
+        Oboe::API.log_exception(nil, e) if log_rails_error?(e)
+        raise
+      ensure
+        Oboe::API.log_exit('rails')
       end
 
       def perform_action(*arguments)
@@ -106,14 +195,19 @@ if defined?(ActionController::Base) && Oboe::Config[:action_controller][:enabled
       end
 
       def rescue_action(exn)
-        Oboe::API.log_exception(nil, exn)
+        Oboe::API.log_exception(nil, exn) if log_rails_error?(exn)
         rescue_action_without_oboe(exn)
       end
 
       def render(options = nil, extra_options = {}, &block)
-        Oboe::API.trace('actionview', {}) do
-          render_without_oboe(options, extra_options, &block)
-        end
+        Oboe::API.log_entry('actionview')
+        render_without_oboe(options, extra_options, &block)
+
+      rescue Exception => e
+        Oboe::API.log_exception(nil, e) if log_rails_error?(e)
+        raise
+      ensure
+        Oboe::API.log_exit('actionview')
       end
     end
   end